Issue 20, 21. Implement single-cluster 2PL.
Note that we currently don't discriminate between read and write locks.
This is still to be done, but the basic functionality doesn't require it.

Further, we don't actually replicate to replicas (hence why it's
single-cluster), this too is forthcoming.

As an aside, planning on forcing reads to wait if there is a write in the
lock queue. Avoids the problem of lots of reads indefinitely delaying
writes.