Adds a transaction manager.
Provides a layer of indirection for clients to communicate with the
transaction manager (TM).

Several important design semantics are embedded into this change:
	- The user sends the entire transaction as one block to the TM.
	- The TM executes a small interpreter to allow the user to execute
	  logic. The Thebes Transactional Language (TTL) is specified in
	  short below.
	- The CLI App has a slightly more confusing interface, since we
	  need to potentially send requests that won't be actually run
	  until END TRANSACTION. Further, we need to be able to send
	  arbitrary TTL commands to be run during the transaction.

Thebes Transactional Language
A very simple, stack-based (prefix operator) language. Currently supports
+, -, *, / as expressions and put/get as statements. A small TTL program
might look like
	put x 3
	get y
	put z * x + y 1  # set z to x*(y+1)

Note that all variables are actually stored server-side via the standard
2PL protocol.
This means that in order to use a variable, you must first "get" or
"put" it.
Thus, the above code would not be valid without the "get y" command in it.

This language was implemented such that it should be easy to add new
statements/expressions (or even control constructs, though unadvisable).