Change 2PL to allow Read-Write locks.
There's a lot of concurrency stuff going on in here, so it would be worth
a close look. Note that we now only synchronize on the LockState (which is
the lock state for a single key), instead of the entire lock manager.

Readers queue behind writers, even if read locks are currently out, to
ensure writers don't wait forever in read-heavy workloads.

The TM now also analyzes the transaction to take out either a write lock
or read lock depending on what it will need, since locks cannot (yet) be
upgraded.